Add public variable assigning

+++ src/main/java/xaridar/lscript/interpreting/types/LFloat.java
@@ -7,25 +7,18 @@ package xaridar.lscript.interpreting.types;
  * @author Xaridar
  */
 
-/*
- * LScript is an interpreted scripting language with static typing, written in Java by Xaridar.
- *
- * @version 2.0.0
- * @author Xaridar
- */
-
-
+/**/
 public class LFloat extends LNumber {
 
     private final float value;
 
     public LFloat(float value) {
-        super("float", Float.valueOf(value));
+        super("float", value);
         this.value = value;
     }
 
     public LFloat(int value) {
-        super("float", Float.valueOf(value));
+        super("float", (float) value);
         this.value = value;
     }
